Change ICompressor implement Compress for a Stream
