+++ ImageCore/Core/ICompressor.cs
@@ -1,4 +1,6 @@
-﻿namespace Image.Core
+﻿using System.IO;
+
+namespace Image.Core
 {
     /// <summary>
     ///     ICompressor is an interface for implementing image compressors.
@@ -10,5 +12,11 @@
         /// </summary>
         /// <param name="fileName">The file name of the image to be compressed.</param>
         void Compress(string fileName);
+
+        /// <summary>
+        ///     The method compresses an image in place.
+        /// </summary>
+        /// <param name="stream">The stream of the image to be compressed.</param>
+        void Compress(Stream stream);
     }
 }
